tests update

+++ challenges/tuple-structs/src/lib.rs
@@ -1,5 +1,5 @@
-pub struct Rectangle(pub u32, pub u32);
+pub struct Rectangle(pub f32, pub f32);
 
-pub fn area(rect: &Rectangle) -> u32 {
+pub fn area(rect: &Rectangle) -> f32 {
     rect.0 * rect.1
 }
